# Mejora rendimiento: Optimización web (76% móvil, 80% escritorio)