--- a/src/components/sections/Skills/Skills.tsx
+++ b/src/components/sections/Skills/Skills.tsx
@@ -2,13 +2,20 @@ import { useSkills } from '../../../hooks';
 import { useTheme } from '../../../context';
 import { DraggableTechCard } from '../../common/TechCard';
 import { motion } from 'framer-motion';
+import { useEffect, useState } from 'react';
 
 const Skills = () => {
   const { skills, loading: skillsLoading } = useSkills();
   const { darkMode } = useTheme();
+  const [mounted, setMounted] = useState(false);
   
   // Combinar todas las tecnologías en un solo array con verificación null/undefined
   const allTechItems = skills?.flatMap(category => category?.items || []) || [];
+
+  // Establecer el estado de montado después de que el componente se monte
+  useEffect(() => {
+    setMounted(true);
+  }, []);
 
   return (
     <section id="skills" className="py-16 relative">
@@ -26,33 +33,36 @@ const Skills = () => {
           </h2>
         </motion.div>
 
-        {skillsLoading ? (
-          <div className="flex justify-center items-center h-[400px]">
-            <div className="animate-spin rounded-full h-16 w-16 border-t-2 border-b-2 border-blue-500"></div>
-          </div>
-        ) : (
-          <div className="relative">
-            {/* Grid directo de tecnologías */}
-            {allTechItems.length > 0 ? (
-              <div className="flex flex-wrap justify-center gap-3 py-8 relative z-10">
-                {allTechItems.map((tech) => (
-                  <DraggableTechCard
-                    key={tech?.id_tech || `tech-${Math.random()}`}
-                    id={tech?.id_tech || 0}
-                    name={tech?.name || 'Sin nombre'}
-                    icon={tech?.icon || ''}
-                    url={tech?.url || ''}
-                    darkMode={darkMode}
-                  />
-                ))}
-              </div>
-            ) : (
-              <div className="text-center py-8">
-                <p>No hay tecnologías disponibles</p>
-              </div>
-            )}
-          </div>
-        )}
+        {/* Contenedor con altura mínima fija para evitar cambios de diseño */}
+        <div className="min-h-[400px]">
+          {skillsLoading || !mounted ? (
+            <div className="flex justify-center items-center h-[400px]">
+              <div className="animate-spin rounded-full h-16 w-16 border-t-2 border-b-2 border-blue-500"></div>
+            </div>
+          ) : (
+            <div className="relative">
+              {/* Grid directo de tecnologías */}
+              {allTechItems.length > 0 ? (
+                <div className="flex flex-wrap justify-center gap-3 py-8 relative z-10">
+                  {allTechItems.map((tech) => (
+                    <DraggableTechCard
+                      key={tech?.id_tech || `tech-${Math.random()}`}
+                      id={tech?.id_tech || 0}
+                      name={tech?.name || 'Sin nombre'}
+                      icon={tech?.icon || ''}
+                      url={tech?.url || ''}
+                      darkMode={darkMode}
+                    />
+                  ))}
+                </div>
+              ) : (
+                <div className="text-center py-8">
+                  <p>No hay tecnologías disponibles</p>
+                </div>
+              )}
+            </div>
+          )}
+        </div>
       </div>
       
       {/* Separador visual inferior */}