--- a/src/index.css
+++ b/src/index.css
@@ -22,3 +22,31 @@ html, body {
   max-width: 100vw;
   overflow-x: hidden;
 }
+
+/* Animación de flotación para las tarjetas de tecnología */
+@keyframes float {
+  0% {
+    transform: translateY(0px);
+  }
+  50% {
+    transform: translateY(-5px);
+  }
+  100% {
+    transform: translateY(0px);
+  }
+}
+
+.animate-float {
+  animation-name: float;
+  animation-timing-function: ease-in-out;
+  animation-iteration-count: infinite;
+}
+
+/* Sombras personalizadas para dark/light mode */
+.hover\:shadow-dark:hover {
+  box-shadow: 0px 15px 30px rgba(0, 0, 0, 0.25);
+}
+
+.hover\:shadow-light:hover {
+  box-shadow: 0px 15px 30px rgba(255, 255, 255, 0.15);
+}