--- a/vite.config.ts
+++ b/vite.config.ts
@@ -2,20 +2,18 @@ import { defineConfig } from 'vite'
 import tailwindcss from '@tailwindcss/vite'
 import react from '@vitejs/plugin-react'
 import { splitVendorChunkPlugin } from 'vite'
-import { visualizer } from 'rollup-plugin-visualizer'
+// Eliminada la importación de visualizer
 
+// https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
     react(),
     tailwindcss(),
     splitVendorChunkPlugin(),
-    visualizer({
-      open: false,
-      gzipSize: true,
-      brotliSize: true,
-    }),
+    // Eliminado el plugin visualizer que generaba stats.html
   ],
   build: {
+    assetsInlineLimit: 1024, // Solo inlinear assets muy pequeños (<1KB)
     chunkSizeWarningLimit: 800,
     reportCompressedSize: true,
     minify: 'terser',
@@ -29,14 +27,16 @@ export default defineConfig({
       output: {
         manualChunks: {
           'react-vendor': ['react', 'react-dom', 'react-router-dom'],
-          'ui-libs': ['framer-motion', 'styled-components'],
+          'framer-motion': ['framer-motion'],
+          'styled-components': ['styled-components'],
           '3d-libs': ['three', '@react-three/drei', '@react-three/fiber'],
           'icons': ['react-icons', 'devicons-react', 'simple-icons'],
-        },
+        }
       },
     },
   },
   optimizeDeps: {
-    include: ['react', 'react-dom', 'react-router-dom', 'framer-motion'],
+    include: ['react', 'react-dom', 'react-router-dom'],
+    exclude: ['framer-motion']
   },
 })