… src/assets/images/logo/optimized/JDSnoppyLogo-small.webp
… src/assets/images/logo/optimized/JDSnoppyLogoHandsUp-small.webp