- Reducción de CLS: Contenedores con altura fija en Skills y Footer
- Optimización de imágenes: Versiones más pequeñas para los logos
- Reducción de JavaScript: DraggableTechCard con CSS nativo en lugar de framer-motion
- Configuración Vite: Mejor separación de chunks para reducir carga inicial
- Mejora tiempos de carga: FCP 1.5s (antes 11s), LCP 1.9s (antes 12s)
- Eliminación de plugin visualizer para evitar generación de stats.html